ResumeGenerator/requirements.txt
pyyaml
pdflatex